refactor(ui): move deep-fryer components into ui
